Install css plugin for babel, restructure files where index only renders, app.jsx holds reviews, tests in progress
