+++ babel.config.json
@@ -1,4 +1,10 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
-  "plugins": ["@babel/plugin-transform-runtime", "transform-class-properties"]
+  "plugins": ["@babel/plugin-transform-runtime", "transform-class-properties", [
+    "react-css-modules",
+    {
+      "generateScopedName": "[local]"
+    }
+  ]
+  ]
 }
