+++ server/index.js
@@ -13,12 +13,13 @@ app.use(express.static(`${__dirname}/../react-client/dist`));
 
 app.get('/api/reviews/:productID', (req, res) => {
   const query = req.params.productID;
-
+  console.log(req.params);
   findReview(query, (err, result) => {
     if (err) {
+      res.status(500);
       throw (err);
     } else {
-      res.send(result);
+      res.status(200).send(result);
     }
   });
 });
