+++ test/client_spec/helpers.js
@@ -1,5 +1,4 @@
 /* eslint-disable object-curly-newline */
-
 import expect from 'chai';
 import { mount, render, shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
